+++ packages/profile-sync-controller/src/controllers/user-storage/__fixtures__/mockMessenger.ts
@@ -1,5 +1,5 @@
 import type { NotNamespacedBy } from '@metamask/base-controller';
-import { ControllerMessenger } from '@metamask/base-controller';
+import { Messenger } from '@metamask/base-controller';
 
 import { MOCK_STORAGE_KEY_SIGNATURE } from '.';
 import type {
@@ -38,6 +38,7 @@ type ExternalEvents = NotNamespacedBy<
 
 /**
  * creates a custom user storage messenger, in case tests need different permissions
+ *
  * @param props - overrides
  * @param props.overrideEvents - override events
  * @returns base messenger, and messenger. You can pass this into the mocks below to mock messenger calls
@@ -45,10 +46,7 @@ type ExternalEvents = NotNamespacedBy<
 export function createCustomUserStorageMessenger(props?: {
   overrideEvents?: ExternalEvents[];
 }) {
-  const baseMessenger = new ControllerMessenger<
-    AllowedActions,
-    AllowedEvents
-  >();
+  const baseMessenger = new Messenger<AllowedActions, AllowedEvents>();
   const messenger = baseMessenger.getRestricted({
     name: 'UserStorageController',
     allowedActions: [
@@ -85,12 +83,13 @@ export function createCustomUserStorageMessenger(props?: {
 }
 
 type OverrideMessengers = {
-  baseMessenger: ControllerMessenger<AllowedActions, AllowedEvents>;
+  baseMessenger: Messenger<AllowedActions, AllowedEvents>;
   messenger: UserStorageControllerMessenger;
 };
 
 /**
  * Jest Mock Utility to generate a mock User Storage Messenger
+ *
  * @param overrideMessengers - override messengers if need to modify the underlying permissions
  * @returns series of mocks to actions that can be called
  */